level-02-02 finish

+++ src/main/java/calculator/Calculator.java
@@ -1,9 +1,9 @@
 package calculator;
 
 public class Calculator {
-    int total;
+    double total;
 
-    public int calculate(int num1, int num2, char ch) throws ArithmeticException{
+    public double calculate(int num1, int num2, char ch) throws ArithmeticException{
 
 
 
@@ -15,7 +15,7 @@ public class Calculator {
                 total = num1 * num2;
             } else if(ch == '/'){
                 try{
-                    total = num1 / num2;
+                    total = (double) num1 / num2;
                 } catch (ArithmeticException er){
                     System.out.println("0으로 나눌 수 없습니다");
                     System.out.println("0이 출력됩니다. 제거해주세요");
